#43 Implement code refactoring as per PR review comments and suggestions
