+++ client-app/src/Modals/DonatedItemModal.ts
@@ -1,6 +1,6 @@
 import { Donor } from '../Modals/DonorModal';
 import { Program } from '../Modals/ProgramModal';
-import { DonatedItemStatus } from '../Modals/DonatedItemStatus';
+import { DonatedItemStatus } from './DonatedItemStatusModal';
 
 export interface DonatedItem {
     id: number;
